Use final matching rule when determining if a command is permitted
Before, a command was permitted only if there were no other
rules denying it *anywhere* in the file. This is inconsistent
with the original doas.

As a side effect of the changes, this commit fixes another
inconsistency where the original doas did not allow
using the `nopass` option with `deny` rules, but rsudoas did.